+++ src/pages/ProjectDetail.tsx
@@ -46,6 +46,7 @@ const ProjectDetail: React.FC = () => {
         url={`/projects/${project.id}`}
         image={project.cover}
         type="article"
+        kind="project"
       />
       <div className="container mx-auto px-4 sm:px-6 lg:px-8 max-w-4xl">
         {/* Back Link */}