+++ vite.config.ts
@@ -4,6 +4,9 @@ import tsconfigPaths from "vite-tsconfig-paths";
 
 // https://vite.dev/config/
 export default defineConfig({
+  ssr: {
+    noExternal: ['react-helmet-async'],
+  },
   build: {
     sourcemap: 'hidden',
   },